requirements.txt
asgiref==3.4.1
astroid==2.6.6
boto3==1.20.0
botocore==1.23.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.6
django-allauth==0.41.0
django-appconf==1.0.5
django-compressor==2.4.1
django-compressor-toolkit==0.6.0
django-countries==7.2.1
django-crispy-forms==1.13.0
django-storages==1.12.3
django-tempus-dominus==5.1.2.16
gunicorn==20.1.0
idna==3.2
isort==5.9.3
jmespath==0.10.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
oauthlib==3.1.1
Pillow==8.3.2
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
pylint==2.9.6
pylint-django==2.4.4
pylint-plugin-utils==0.6
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
rcssmin==1.0.6
requests==2.26.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
stripe==2.61.0
toml==0.10.2
urllib3==1.26.7
wrapt==1.12.1
